Handle call widget redirects (#19)
Issue: Upon navigation, the calling widget and the host page would gets out of sync.  There are two distinct issues that need to be resolved - 1) Notifying the host page of the URL change so that the host page could re-bootstrap sending messages to a iFrame with that URL.  2) Send the host page url to call widget so that it can send messages back.  IFrameHost relied on the referer url to find the host page URL, that is no longer the case as the iFrame is triggering the redirect.

Solution - 1)  iFrame should sends  a message to the host page with their new URL - this allows the host page to send messages targeted to an iFrame with that URL  2) the iFrameManager re-bootstrap on page loads and share their URLs through the SYNC and SYNC_ACK messages.

Added a new event (type: SET_WIDGET_URL) for the iFrame to communicate the new URL back to the host page.  

Before redirecting the iFrame to a new URL, send a message as following:
cti.sendMessage({
   type: Constants.messageType.SET_WIDGET_URL,
   iFrameUrl: "https://127.0.0.1:9025"
});

Re: https://github.com/HubSpot/calling-extensions-sdk/issues/18