feat - Added an event handler for receiving a HubSpot created engagement, (#27)
Details -
The  call engagements should have context related associations.  For example, the engagement for a call made to a contact from a deal page should have tha deal associated.  Precalculating these associations requires API calls and sending them as part of dial number payload adds time to the dial number path.  Getting the engagementId from the widget is also not practicle as widget may be creatig the engagements as a batch process given that creating engagement is not critical to connecting a call.

The change allows the widget to receive an engagement created by HubSpot - the engagement will be created on receiving the outgoing_call event and send it back to the widget as part of onEngagementCreated event.  The widget backend can update this engagement with call details instread of creating a new engagement.